requirements.txt
# speech to text
pyaudio
speechrecognition

# openai chat response
python-dotenv
openai

# text to speech
pyttsx3
pygame